refactor: introduce `snow/networking/sender/sendertest`
