+++ snow/networking/sender/sender_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package sender
+package sender_test
 
 import (
 	"context"
@@ -22,6 +22,7 @@ import (
 	"github.com/ava-labs/avalanchego/snow/networking/benchlist"
 	"github.com/ava-labs/avalanchego/snow/networking/handler"
 	"github.com/ava-labs/avalanchego/snow/networking/router"
+	"github.com/ava-labs/avalanchego/snow/networking/sender/sendertest"
 	"github.com/ava-labs/avalanchego/snow/networking/timeout"
 	"github.com/ava-labs/avalanchego/snow/networking/tracker"
 	"github.com/ava-labs/avalanchego/snow/snowtest"
@@ -37,6 +38,8 @@ import (
 
 	p2ppb "github.com/ava-labs/avalanchego/proto/pb/p2p"
 	commontracker "github.com/ava-labs/avalanchego/snow/engine/common/tracker"
+
+	. "github.com/ava-labs/avalanchego/snow/networking/sender"
 )
 
 const testThreadPoolSize = 2
@@ -88,7 +91,7 @@ func TestTimeout(t *testing.T) {
 		prometheus.NewRegistry(),
 	))
 
-	externalSender := &ExternalSenderTest{TB: t}
+	externalSender := &sendertest.ExternalSenderTest{TB: t}
 	externalSender.Default(false)
 
 	sender, err := New(
@@ -365,7 +368,7 @@ func TestReliableMessages(t *testing.T) {
 		prometheus.NewRegistry(),
 	))
 
-	externalSender := &ExternalSenderTest{TB: t}
+	externalSender := &sendertest.ExternalSenderTest{TB: t}
 	externalSender.Default(false)
 
 	sender, err := New(
@@ -522,7 +525,7 @@ func TestReliableMessagesToMyself(t *testing.T) {
 		prometheus.NewRegistry(),
 	))
 
-	externalSender := &ExternalSenderTest{TB: t}
+	externalSender := &sendertest.ExternalSenderTest{TB: t}
 	externalSender.Default(false)
 
 	sender, err := New(
